backend/requirements.txt
Django==6.0
djangorestframework==3.16.1
markdown==3.10
django-filter==25.2